+++ policyengine_us_data/datasets/puf/puf.py
@@ -9,6 +9,7 @@ from policyengine_us_data.datasets.puf.irs_puf import IRS_PUF_2015
 from policyengine_us_data.utils.uprating import (
     create_policyengine_uprating_factors_table,
 )
+import os
 
 rng = np.random.default_rng(seed=64)
 
@@ -18,6 +19,8 @@ def impute_pension_contributions_to_puf(puf_df):
     from policyengine_us_data.datasets.cps import CPS_2021
 
     cps = Microsimulation(dataset=CPS_2021)
+    if os.environ.get("TEST_LITE"):
+        cps.subsample(1_000)
     cps_df = cps.calculate_dataframe(
         ["employment_income", "household_weight", "pre_tax_contributions"]
     )
@@ -45,6 +48,11 @@ def impute_missing_demographics(
         .merge(demographics, on="RECID")
         .fillna(0)
     )
+
+    if os.environ.get("TEST_LITE"):
+        puf_with_demographics = puf_with_demographics.sample(
+            n=1_000, random_state=0
+        )
 
     DEMOGRAPHIC_VARIABLES = [
         "AGEDP1",
@@ -204,7 +212,9 @@ def preprocess_puf(puf: pd.DataFrame) -> pd.DataFrame:
     # Ignore cmbtp (estimate of AMT income not in AGI)
     # Ignore k1bx14s and k1bx14p (partner self-employment income included in partnership and S-corp income)
     qbi = np.maximum(0, puf.E00900 + puf.E26270 + puf.E02100 + puf.E27200)
-    W2_WAGES_SCALE = 0.16
+    # 10.1% passthrough rate for W2 wages hits the JCT tax expenditure target for QBID
+    # https://gist.github.com/nikhilwoodruff/262c80b8b17935d6fb8544647143b854
+    W2_WAGES_SCALE = 0.101
     puf["w2_wages_from_qualified_business"] = qbi * W2_WAGES_SCALE
 
     # Remove aggregate records